add a adnuntius.onClose() event to allow closing the web view from javascript
