+++ README.md
@@ -19,7 +19,7 @@ repositories {
 Add a dependency to your build.gradle file:
 
 ```
-implementation 'com.adnuntius.android.sdk:1.4.0'
+implementation 'com.adnuntius.android.sdk:1.4.1'
 ```
 
 ## Ad Delivery
@@ -71,6 +71,27 @@ A basic api for simple ad integrations, uses adn.js internally to render the ad 
                 }
             });
     }
+```
+
+#### Close View from Layout
+
+Its now possible to close a web view from an adnuntius layout via javascript.
+
+If you want to be able to close ad view from javascript, its not possible to close the parent window from a layout due to the dreaded `Scripts may close only the windows that were opened by them.`   So we have added a new method you can call from javascript which provides this functionality:
+
+```javascript
+if (typeof parent.androidAdnuntius != "undefined") {
+ parent.androidAdnuntius.closeView();
+}
+```
+
+And the corresponding method in the CompletionHandler:
+
+```java
+   @Override
+   public void onClose() {
+        finish();
+   }
 ```
 
 #### Live Preview support
@@ -190,7 +211,7 @@ dataClient.page(page, new DataResponseHandler() {
 
 ## Bugs, Issues and Support
 
-This SDK is a work in progress and will be given attention when necessary based on feed back.  You can raise issues on github or via zen desk at https://admin.adnuntius.com
+You can raise issues on github or via Zen Desk at https://admin.adnuntius.com
 
 # License
