+++ sdk/build.gradle
@@ -3,17 +3,17 @@ apply plugin: 'com.android.library'
 apply plugin: 'maven-publish'
 apply plugin: 'signing'
 
-version = "1.4.0"
+version = "1.4.1"
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "29.0.2"
+    compileSdkVersion 29
+    buildToolsVersion "30.0.3"
 
     defaultConfig {
         // https://www.appbrain.com/stats/top-android-sdk-versions
         // anything below lollipop hardly seems worth it
         minSdkVersion 21
-        targetSdkVersion 26
+        targetSdkVersion 29
         versionName version
         versionCode Integer.valueOf(version.toString().replace('.', '').replace('-SNAPSHOT', ''))
